Modify: TopNavBar를 Nav로 변경하고 불필요한 코드 제거
