--- next-app/src/components/common/Layout/Nav/SideNavBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SideMenuBar';
