--- next-app/src/components/common/Layout/Nav/TopNavBar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TopNavBar'
